Added support for specifying interval and hours.
Added support for specifying the shake interval and
the hours at which it's good to shake. Now, instead
of scheduling as a weekly cron job it should be schedule
as an hourly cron job, so that if it gets interrupted
we won't have to wait another week (or whatever the set
interval is now) for the job to be retried. We can also
specify a list of hours during which it is OK to try to
run the job.